c/snap-bootstrap/bootstrap: fix the build on other distros (except debian)
debian needs more work

drive-by: add a doc comment for Run
