+++ cmd/snap-bootstrap/bootstrap/bootstrap_dummy.go
@@ -1,8 +1,8 @@
 // -*- Mode: Go; indent-tabs-mode: t -*-
-// +build !nosecboot
+// +build nosecboot
 
 /*
- * Copyright (C) 2019 Canonical Ltd
+ * Copyright (C) 2019-2020 Canonical Ltd
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 3 as
@@ -20,7 +20,10 @@
 
 package bootstrap
 
-var (
-	EnsureLayoutCompatibility = ensureLayoutCompatibility
-	DeviceFromRole            = deviceFromRole
+import (
+	"fmt"
 )
+
+func Run(gadgetRoot, device string, options Options) error {
+	return fmt.Errorf("build without secboot support")
+}
