+++ cmd/snap-bootstrap/bootstrap/bootstrap_test.go
@@ -1,4 +1,5 @@
 // -*- Mode: Go; indent-tabs-mode: t -*-
+// +build !nosecboot
 
 /*
  * Copyright (C) 2019 Canonical Ltd
